requirements.txt
# 核心依赖
tlsh==4.8.2
PyYAML>=6.0.1
psutil==5.9.5
chardet>=4.0.0

# 数据处理
numpy>=1.20.0
pandas>=1.3.0

# Web API
flask>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0

# 数据库
sqlalchemy>=1.4.0

# 测试工具
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
pytest-benchmark>=4.0.0
pytest-mock>=3.11.0
pytest-timeout>=2.1.0
pytest-randomly>=3.13.0
coverage>=7.3.0
codecov>=2.1.0
html-testRunner==1.2.1

# 代码质量
flake8>=6.1.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
pylint>=2.17.0
bandit>=1.7.0
safety>=2.3.0

# 文档工具
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0
sphinx-copybutton>=0.5.0

# 开发和调试工具
ipython>=8.14.0
ipdb>=0.13.0
debugpy>=1.6.0
build>=1.0.0
twine>=4.0.0
wheel>=0.41.0

# 性能分析
memory_profiler>=0.61.0
line_profiler>=4.1.0
py-spy>=0.3.0

# 类型检查
types-PyYAML>=6.0.12.12
types-psutil==5.9.5.17
types-requests>=2.31.0
types-setuptools>=68.0.0

# 代码解析
javalang==0.13.0
libclang==16.0.6

# 代码克隆检测
scikit-learn>=1.0.0
gensim>=4.0.0
nltk>=3.6.0

# 版本预测
scipy>=1.7.0
statsmodels>=0.13.0
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0 